refactor: split link-tree from sidebar modal
